bionic_backend/requirements.txt
asgiref==3.11.1
certifi==2026.2.25
charset-normalizer==3.4.6
dj-database-url==3.1.2
Django==6.0.3
django-cors-headers==4.9.0
djangorestframework==3.17.1
djangorestframework_simplejwt==5.5.1
gunicorn==25.3.0
idna==3.11
mysql-connector-python==9.6.0
packaging==26.0
psycopg2-binary==2.9.11
PyJWT==2.12.1
requests==2.33.0
sqlparse==0.5.5
tzdata==2025.3
urllib3==2.6.3
